Added increment quantity buttons for non-unique heroes that are following another hero.
